also export colors to JSON
