+++ packages/themes/.gitignore
@@ -1,1 +1,3 @@
 colors.css
+colors.json
+
